authorize memberships_controller

+++ app/controllers/memberships_controller.rb
@@ -1,19 +1,21 @@
 class MembershipsController < ApplicationController
   def index
-    # TODO: authorize
     respond_to do |format|
       format.json do
         @room = Room.includes(project: :users).find(params[:room_id])
+        authorize(@room, policy_class: MembershipPolicy)
+
         @memberships = @room.memberships.includes(user: :profile)
       end
     end
   end
 
   def destroy
-    # TODO: authorize
     respond_to do |format|
       format.json do
         membership = Membership.find(params[:id])
+        authorize(membership.room, policy_class: MembershipPolicy)
+
         if membership.destroy
           membership.broadcast_membership(Membership::REMOVE_MEMBERSHIP)
           @room = membership.room
@@ -27,10 +29,11 @@ class MembershipsController < ApplicationController
   end
 
   def create
-    # TODO: authorize
     respond_to do |format|
       format.json do
         membership = Membership.new(membership_params)
+        authorize(membership.room, policy_class: MembershipPolicy)
+
         if membership.save
           membership.broadcast_membership(Membership::ADD_MEMBERSHIP)
           @room = membership.room
